Feat: fixed appointments
