+++ HospitalApp/HospitalApp/Models/Doctors.cs
@@ -1,0 +1,7 @@
+public class Doctor
+{
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public string Specialty { get; set; }
+    public string Contact { get; set; }
+}
