+++ HospitalApp/HospitalApp/Views/DashboardPageView.axaml.cs
@@ -1,13 +1,21 @@
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
+using HospitalApp.ViewModels;
 
-namespace HospitalApp.Views;
-
-public partial class DashboardPageView : UserControl
+namespace HospitalApp.Views
 {
-    public DashboardPageView()
+    public partial class DashboardPageView : UserControl
     {
-        InitializeComponent();
+        public DashboardPageView()
+        {
+            InitializeComponent();
+            DataContext = new DashboardPageViewModel();
+        }
+
+        private void InitializeComponent()
+        {
+            AvaloniaXamlLoader.Load(this);
+        }
     }
 }
